requirements.txt
torch==2.3.1+cu121
torchvision==0.18.1+cu121
torchaudio==2.3.1+cu121
pytorch-lightning>=2.0.0
tensorboard
captcha
pyyaml
pillow
numpy<2
matplotlib
opencv-python>=4.5.0
pandas>=1.3.0
colorama>=0.4.0
wheel>=0.38.0 # not directly required, pinned by Snyk to avoid a vulnerability